+++ kubernetes/worker-controller/entrypoint.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+SD=$(dirname $0)
+
+exec $SD/worker_controller.py